Modularize public navbar and align list editor
Refs #354
