+++ web/src/components/design-system/index.ts
@@ -6,6 +6,7 @@ export { NextActionStrip } from './next-action-strip';
 export type { NextActionStripProps } from './next-action-strip';
 export { PriceRow } from './price-row';
 export type { PriceRowProps } from './price-row';
+export { PublicNavbar } from './public-navbar';
 export { StatusBadge, statusBadgeVariants, statusPresets } from './status-badge';
 export type { StatusBadgeProps, StatusFamily } from './status-badge';
 export { StickyActionBar } from './sticky-action-bar';
